Implemented Game Timer and Changed class name
I implemented the GameTimer and changed the class name
to MainGUIApplication. For some reason the name Main was 
causing issues when compiling in command line. Also I had 
to remove the package statement at the beginning of the 
file, and rennamed the t variable speedAdjustmentFactor and
added it to the instance variables. Should work with the current
files.