Created new game folder
This version of the game reverts Obstacles and Collectibles to child classes of Items, and
incorporates a highly truncated version of AnimationApplication.java into MainGUIApplication, that
only involve Obstacles and Collectibles. The MainGUIApplication is now fully OO.

Methods for moveUp() and moveDown() were also added to Player to improve encapsulation in 
JumpEvent.java. 

A very small update to GameTimer.

Game now also uses the original method from TextOutput to output current active objects
to the console each frame, with some slight adjustments to the aforementioned method. 

Also added a class diagram.